fix time conplexity

+++ ex11/test_matrix.cpp
@@ -37,7 +37,7 @@ void testMatrixREFValid(UnitTest &test) {
         ASSERT_EQUALS(test, 1032.0, u.determinant());
     }
     {
-        Matrix<int> u({
+        Matrix<float> u({
             {1, 2, 3, 4, 5, 6, 7, 8, 9},
             {11, 12, 13, 14, 15, 16, 17, 18, 19},
             {21, 22, 23, 24, 25, 26, 27, 28, 29},
